pyproject.toml
[tool.poetry]
name = "app_rag_chat_bot"
version = "0.1.0"
description = "RAG Chatbot con Memoria Dinámica"
authors = ["Javier Lazaro Garcia <your-email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.20"
langchain-openai = "^0.3.8"
langchain-community = "^0.3.19"
openai = "^1.66.3"
chromadb = "0.4.18"
pypdf = "3.17.0"
gradio = "4.44.1"
pydantic = "^2.10.6"
python-dotenv = "1.0.0"
pydantic-settings = "^2.8.1"
langchain-chroma = "^0.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
